doc: restore: remove note on --patch w/ pathspecs
This note was added with the command in 46e91b663b (checkout: split part of
it to new command 'restore', 2019-04-25), but it is now inaccurate. The
underlying builtin `add -i` implementation, made default in 0527ccb1b5 (add
-i: default to the built-in implementation, 2021-11-30), supports pathspecs,
so `git restore -p <pathspec>...` has worked for all users since then. I
bisected to verify this was the commit that added support.

Signed-off-by: Adam Johnson <me@adamj.eu>