docs: add git hash-object -t option's possible values
The verbiage under the NAME section for git hash-object could
lead one to conclude that git hash-object can only be used to create
blobs when in fact the description makes it clear that it can be used to
create objects, not just blobs. Lets clarify this in the NAME text.

Further, the description for the option -t does not list out other types
that can be used. Let's make this explicit by listing out the different
object types.

Signed-off-by: John Cai <johncai86@gmail.com>