docs: add git-hash-object -t option's possible values
For newer users of Git, the possible values of -t in git-hash-object may
not be apparent. In fact the current verbiage under NAME could
lead one to conclude that git-hash-object(1) can only be used to create
blobs.

Update the verbiage to make it clear the command can be used to write
objects, not just blobs. Also add the possible values for -t.

Signed-off-by: John Cai <johncai86@gmail.com>