unresolve: make sure resolve-undo info is left even for removed paths
"git rerere forget" requires that resolve-undo information is
correctly recorded in the index.

The current rerere machinery is not even prepared to access the
resolve-undo information for paths resolved for removal, as it first
goes to the main index entry and then finds the resolve-undo
information associated to the same pathname.  This has to be fixed
later (the codepath used by "git update-index --unresolve" shares
the same problem---unmerge_index_entry_at() is a flawed interface
that assumes that the index has the path in it after resolution, and
cannot be used when the resolution was to remove), but let's first
make sure we have the information recorded.

Signed-off-by: Junio C Hamano <gitster@pobox.com>