checkout -m: unmatching pathspec that matches only removals is OK
The "git checkout [-m] <pathspec>" codepath makes sure that all
elements of <pathspec> match what is in the index by walking the
main cache_entry array of the current index, and dies when there
are elements that did not match.  This is to prevent mistakes like

    $ git checkout RAEDME

that silently turns into a no-op, which is logical (we ask "check
out all the paths that match this pathspec" and check out 0 paths)
but unexpected.

That is fine for checking out the paths that are in the index, but
when recreating the conflicted state with "git checkout -m", the
request may ask to work with a path that the user has resolved for
removal (e.g. "they modified while we removed" can be resolved to
removal).  In such a case, the path that used to be in the
conflicted state cannot be found in the main cache_entry array.

The resolve_undo information, that is used to resurrect conflicted
state, is stored even for such a removed path.  We should check there
to see if given pathspec matches to avoid false errors.

Note that this step alone does not allow "checkout -m" to resurrect
"they modified while we removed" conflict yet.  The API function
unmerge_marked_index() that is used to unresolve formerly conflicted
paths (incorrectly) assumes that it can use the offset in the main
cache_entry_array of the index to specify which path to unresolve,
which obviously does not work once a conflicted path is resolved for
removal.  It needs to be fixed but in a separate step.

The early "compute pathspec matches once and reuse the information
by marking cache entries with CE_MATCHED bit" came from e721c154
(checkout: avoid unnecessary match_pathspec calls, 2013-03-27).
Eventually the "unmerge" codepath needs to be separated from it,
because the pathspec matches cannot always be recorded in the cache
entries (i.e. removed paths do not have any cache entries in the
index).  That is also a topic of future steps.

Signed-off-by: Junio C Hamano <gitster@pobox.com>